+++ midilib/src/main/cpp/player/WavePlayer.cpp
@@ -1,7 +1,7 @@
 #include "WavePlayer.h"
 #include "../AudioEngine.h"
 
-void WavePlayer::fillBuffer(float *buffer, int32_t numFrames) {
+void WavePlayer::fillBuffer(float *buffer, size_t numFrames) {
     for (int32_t i = 0; i < numFrames; i++) {
         float sampleValue = 0;
         for (auto & channel : AudioEngine::getChannels()) {
